ODA_project/requirements.txt
Django==4.0.4
black==22.3.0
django-admin-interface==0.19.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mysqlclient == 2.1.0
requests == 2.27.1
drf-yasg == 1.20.0